requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
apscheduler=3.10.4=pypi_0
asttokens=2.4.1=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2023.12.12=h06a4308_0
certifi=2024.2.2=py310h06a4308_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
contourpy=1.2.0=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.99=0
cuda-runtime=12.1.0=0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
docopt=0.6.2=pypi_0
et-xmlfile=1.1.0=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.13.1=pypi_0
flask=3.1.0=pypi_0
fonttools=4.50.0=pypi_0
fsspec=2024.3.0=pypi_0
googleapis-common-protos=1.63.2=pypi_0
grpcio=1.64.1=pypi_0
grpcio-tools=1.64.1=pypi_0
idna=3.6=pypi_0
ipython=8.12.3=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pypi_0
jetpack=1.1=pypi_0
jinja2=3.1.3=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
kiwisolver=1.4.5=pypi_0
lapx=0.5.9.post1=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.0.20=0
libcurand=10.3.5.119=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
loguru=0.7.2=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.3=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.99=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python=4.9.0.80=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=24.0=pypi_0
pandas=2.2.1=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py310h06a4308_0
pipreqs=0.5.0=pypi_0
platformdirs=4.2.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=5.28.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-cpuinfo=9.0.0=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.2=pypi_0
pystache=0.6.5=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytorch-cuda=12.1=ha16c6d3_5
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.34.0=pypi_0
requests=2.31.0=pypi_0
rpds-py=0.18.0=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypi_0
thop=0.1.1-2209072238=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.2.1=pypi_0
torchaudio=2.2.1=pypi_0
torchvision=0.17.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
tzlocal=5.2=pypi_0
ultralytics=8.1.29=pypi_0
urllib3=2.2.1=pypi_0
wcwidth=0.2.13=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.6=h5eee18b_0
yarg=0.1.9=pypi_0
zlib=1.2.13=h5eee18b_0
